requirements.txt
Django==5.2.6
djangorestframework==3.15.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8
mysqlclient==2.2.0